pyproject.toml
[project]
name = "crisis-bench"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "matplotlib>=3.10.7",
    "numpy>=2.3.5",
    "openai>=2.8.1",
    "pydantic>=2.12.4",
    "pyyaml>=6.0.3",
    "tabulate>=0.9.0",
]
